+++ src/main/kotlin/org/jetbrains/lama/psi/elementTypes/LamaElementFactory.kt
@@ -2,6 +2,7 @@ package org.jetbrains.lama.psi.elementTypes
 
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFileFactory
+import com.intellij.psi.PsiWhiteSpace
 import org.jetbrains.lama.psi.LamaLanguage
 import org.jetbrains.lama.psi.api.LamaFile
 import org.jetbrains.lama.psi.api.LamaPsiElement
@@ -10,6 +11,11 @@ object LamaElementFactory {
   fun createLamaPsiElementFromText(project: Project, text: String): LamaPsiElement {
     val file = buildLamaFileFromText(project, text)
     return file.firstChild as LamaPsiElement
+  }
+
+  fun createNl(project: Project): PsiWhiteSpace {
+    val file = buildLamaFileFromText(project, "\n")
+    return file.firstChild as PsiWhiteSpace
   }
 
   fun buildLamaFileFromText(project: Project, text: String): LamaFile {